Level 2 tweaks, Health system more set up, and Anchor points fixed
I got the exploding Ai to effect the health bar ( 1bug takes 0.25/1 health). The jellies add to the health bar, the anchor points on the main menu as well as the hud (all 3 levles and health bar). also moved some stuff on level 2 trying to make the jumps a little more difficult.